+++ casper/src/main/scala/coop/rchain/casper/util/rholang/CasperRhoRuntimeSyntax.scala
@@ -51,7 +51,8 @@ import coop.rchain.rholang.interpreter.errors.BugFoundError
 import coop.rchain.rholang.interpreter.{EvaluateResult, ReplayRhoRuntime, RhoRuntime}
 import coop.rchain.rspace.hashing.Blake2b256Hash
 import coop.rchain.rspace.history.History.emptyRootHash
-import coop.rchain.rspace.{trace, ReplayException}
+import coop.rchain.rspace.trace
+import coop.rchain.rspace.util.ReplayException
 import coop.rchain.shared.Log
 
 trait RhoRuntimeSyntax {
