+++ casper/src/main/scala/coop/rchain/casper/util/rholang/ReplayFailure.scala
@@ -2,7 +2,7 @@ package coop.rchain.casper.util.rholang
 
 import coop.rchain.casper.protocol.DeployData
 import coop.rchain.crypto.signatures.Signed
-import coop.rchain.rspace.ReplayException
+import coop.rchain.rspace.util.ReplayException
 
 sealed trait ReplayFailure
