+++ rspace/src/main/scala/coop/rchain/rspace/util/ReplayException.scala
@@ -1,3 +1,3 @@
-package coop.rchain.rspace
+package coop.rchain.rspace.util
 
 class ReplayException(msg: String) extends Exception(msg)
